requirements.txt
httpx>=0.27.0
loguru>=0.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
tenacity>=8.0.0
openai>=1.0.0
apscheduler>=3.10.0
